update intro_test.dart, welcome.dart, and intro.dart
Starting off the update intro.dart has had an dialogue prompt added that offers the tutorial or the option to skip it. The intended effects are to either start the tutoral or skip to the welcome page which should have the option to starting a fresh account. The tutorial page is not yet made so for now it goes to the dashboard page and skipping goes to the welcome page.

In welcome.dart I edited the size of the logo to make it a more approriate size.

Laslty in intro_test.dart  I added a case for each button option so for start tutorial we expect the dashboard page and for skip we expect the welcome page.