updated welcome_test.dart and welcome.dart
updated welcome.dart to have the app bar removed since there was no purpose(it was transparent), and replaced the buttons previously used to hold the title and welcome text with container widgets. The test was also updated to check for a element in the body as appose to the app bar.