update welcome_test.dart and welcome.dart
the test have been created in welcome_test.dart that now accomidate for a shared preference variable. When true we expect the password field to appear and when false we expect the start button to appear. Currently both test fail but that is because the only code added to welcome was the import line.