update to welcome, welcome_test, and pubspec
the update to pubspec was made to move shared preferenced to the current dependencies branch.

Welcome_test has been modified to add in mock values for shared preferences and both test are in a working state. Work is needed to fix a new bug caused when the logo is full size making a new test fail.

Welcome has the added functionality of shared preferences that are currently in a semi functional state. The app will update the shared preference to open the password prompt but after closing the app and reopening it the shared preference is not saved so it is reset. other than the main functionality not working all other code has been added to display a start button.