update to welcome.dart and welcome_test.dart
The shared preferences now appear to be working as intended and all the created test cases pass. I am going to add a couple more since code coverage is low with 2 added function. Other than that everything appears to work but functionality will also be worked on as the shared preferences do not clear on restart.