update to welcome.dart and welcome_test.dart
For welcome_test I added in an extra test case to consider when the shared preferences are yet to be initialized and added in a line to test the onpressed function within the start button.

For welcome I added in a setState function to update the screen after the shared preference is updated and I added in a line of code to the reset button that can be commented out to test if the shared preferences are being updated in real time. Currently I can close the app and reopen it and the shared preference is saved so I am going to focus on fixing the error caused when scaling up the logo in the welcome screen and this function should be just about complete.