Update to welcome.dart
I have the basic logic of the page working with the custom password field made. I also modified the validation checker so it can be used with the new password field but currently things only work if the user chooses to not save a password. The expected behavior is that the start button is there and the prompt to set up a password does not get asked again unless the reset password button is hit which resets all the shared preferences as if the user is opening the app for the first time.