update to fields.dart, welcome.dart , and main.dart
fields.dart has an added field that accounts for when the form is submitted and if the password passes the validator check then the next page is entered(dashboard).

main.dart has just had welcomePage replaced with dashboardPage to fix one failing test.

Welcome.dart has been changed so that we are saving a password if a user decides to use a password and if they choose not to then they will be presented with the start button on open that opens up to the dashboard page. If they use a password there is a password field that validates the password before opening the dashboard.