update to main.dart, welcome.dart, and welcome_test.dart
In main I removed the unused import to fix an error.

In welcome.dart I removed some of the comments and adjusted others to help make the code more readable.

in Welcome_test.dart I modified 3 test to work with the new start button functionality and created 2 additional test to increase coverage. This is the test commit to see the current coverage and to run test on.