update to create test for new tag entry page
Settings_tag_test.dart is the new test file that will be used to test the settings_tag.dart file. This test will ensure that tags can be searched, added, and deleted from the tag page along with confirming the page is accessable from the settings page.

settings_test.dart was modified to move the button test for edit tag list to the end because it now redirects to another page.

Settings.dart was modified to now direct the edit tag list page.