Create LICENSE
Per the requirements of the Senior Design project, we need a license. This one I believe, fits the best with what we are trying to do.

Like all the others there is a grant of patent (means anyone can use it), this one protects the licensing though, which means it cannot be used in close source implementations.

This means that the software, in theory, would always be free or open source.

Please let me know your thoughts.