Update to encryptor, settings, and decorations.dart
This update displays a loading screen during the encryption process for logging into the app and for resetting the password. Some tests will need to be updated but there also needs to be an added loading screen during password creation during the time it takes to save the password. Instead of going to the next page automatically there is a dialog box to go there and this should also be replaced.