update to respond to github comments
this update seeks to fix a text overflow issue that was brought up in the pull request. I also reverted an unnecessary change that was pointed out and adjusted the new timeout to match the actual time it takes for the test to complete. (from 6 to 4 minutes after reducing number of iteration from 9 to 6 during encryption process)