update to replace loading logo and fix graphics issue
This commit seeks to accommodate to the new github review requests made. I have added in a new glasses image that is centered and used that for the loading animation. I also fixed the barrier issues where the backgrounds were getting messed up due to the amount of alerts being displayed.